refactor: use constate to better manage locally global state
