+++ src/pages/ContentScripts/features/repo-collection/view.tsx
@@ -1,33 +1,15 @@
 import { CollectionButton } from './CollectionButton';
 import { CollectionModal } from './CollectionModal';
-import { CollectionContext } from './context';
+import { RepoCollectionProvider } from './context';
 
-import React, { useState } from 'react';
+import React from 'react';
 
 const View = () => {
-  const [hideCollectionList, setHideCollectionList] = useState(false);
-  const [hideAddToCollections, setHideAddToCollections] = useState(true);
-
-  const [showModal, setShowModal] = useState(false);
-  const [selectedCollection, setSelectedCollection] = useState<string>();
-
-  const contextValue = {
-    hideCollectionList,
-    setHideCollectionList,
-    hideAddToCollections,
-    setHideAddToCollections,
-
-    showModal,
-    setShowModal,
-    selectedCollection,
-    setSelectedCollection,
-  };
-
   return (
-    <CollectionContext.Provider value={contextValue}>
+    <RepoCollectionProvider>
       <CollectionButton />
       <CollectionModal />
-    </CollectionContext.Provider>
+    </RepoCollectionProvider>
   );
 };
